+++ R/property-helpers.R
@@ -33,6 +33,7 @@ huxtable_table_attrs <- c(
   "caption",
   "caption_pos",
   "caption_width",
+  "breakable",
   "tabular_environment",
   "table_environment",
   "label",
@@ -64,6 +65,7 @@ huxtable_env$huxtable_default_attrs <- list(
   caption             = NA_character_,
   caption_pos         = "top",
   caption_width       = NA_real_,
+  breakable           = FALSE,
   position            = "center",
   tabular_environment = NA_character_,
   table_environment   = "table",
@@ -124,7 +126,7 @@ validate_prop <- function(value, prop, check_fun = NULL, check_values = NULL,
 #' @param check_fun    Optional validation function.
 #' @param check_values Optional vector of allowed values.
 #' @param extra        Extra code to run after validation.
-#' @param reset_na     Passed to [`validate_prop`].
+#' @param reset_na     Should `NA` values be replaced with the huxtable default?
 #'
 #' @noRd
 prop_set <- function(ht, prop, row, col, value = NULL, fn = NULL,