+++ NEWS.md
@@ -2,10 +2,16 @@
 
 * This is the last update which will be pushed to CRAN. Future versions will be 
   available on GitHub and/or via https://hughjonesd.r-universe.dev.
+* New table property `breakable()` allows tables to break between rows across
+  pages in HTML, LaTeX, RTF, Typst and Word output. Breakable tables
+  are implemented in LaTeX via the `longtable` package, meaning this is a new
+  LaTeX package dependency.
 * Bugfix: `add_rownames()` should no longer add a 1 in the header row of a
   data frame.
 * `as_Workbook()` is now much faster when adding many tables to an Excel
   workbook. Thanks @lemonad for the report.
+* `huxtable_html_css()` is now exported, for including huxtable's default CSS
+  once in an HTML document.
   
 # huxtable 5.8.0
 
@@ -789,5 +795,3 @@ For example, the following code now does what you probably want:
 
 * Added a `NEWS.md` file to track changes to the package.
 * First CRAN release.
-
-